+++ spimdisasm/mips/MipsFileSplits.py
@@ -5,7 +5,6 @@
 
 from __future__ import annotations
 
-from pathlib import Path
 
 from .. import common